Changed the Distance number of each card to byte instead of int

+++ deck/card/Card.java
@@ -2,6 +2,6 @@ package deck.card;
 
 public abstract class Card {
 	public abstract String getAbilityDescription();
-	public abstract int getDistance();
+	public abstract byte getDistance();
 	public abstract String getName();
 }
